Make bottom one-fourth of web page
Make bottom one-fourth of the web page. This includes the 'call to
action' area (which has the second 'sign up' button) and the footer.
Margins and paddings still need to be fixed and readme has to be
updated. Making the web page more alive by replacing the boxes and
texts with actual images and relevant texts is still under
consideration.